refactoring: inject logger into rmq consumer
